foodPath generates successfully for first food. There is some issue with finding the first point, but I put a band-aid over it which seems to be holding for now.
Doesn't move on to next food because when finding direction to travel to next point, it references a future point. Every time the snake moves it removes that move so we end up looking for a point out of bounds before we reset for the next food. Need to fix